vk_project/requirements.txt
alphabet-detector==0.0.7
amqp==2.2.2
anyio==4.2.0
asgiref==3.7.2
atomicwrites==1.1.5
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.5.0.3
celery==4.2.1
cryptography==43.0.0
cffi==1.17.0
certifi==2018.4.16
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
coverage==5.0.4
Django==3.2.23
django-bootstrap4==21.1
django-constance==3.1.0
django-picklefield==3.1
django-pytest==0.2.0
enum34==1.1.6
exceptiongroup==1.2.0
freezegun==1.5.1
gunicorn==19.8.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==2.1.3
iniconfig==1.1.1
kombu==4.2.2.post1
Markdown==3.3.4
more-itertools==4.3.0
numpy==1.24.4
packaging==21.3
pkce==1.0.3
phonenumbers==8.9.5
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pyparsing==3.0.8
pytesseract==0.2.4
pytest==7.1.2
pytest-django==3.8.0
pytest_freezer==0.4.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytube==9.3.0
pytz==2023.4
ratelimit==2.2.0
ratelimitingfilter==1.2
raven==6.9.0
requests==2.31.0
sentry-sdk==0.3.6
six==1.11.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.4
tenacity==5.0.4
tomli==2.0.1
typing_extensions==4.7.1
uritools==2.2.0
urlextract==0.8.3
urllib3==1.22
vine==1.1.4
vk-api==11.9.9
vk-requests==1.2.0
wcwidth==0.1.9
zipp==3.1.0
